fixed #ioc-332
UsingFactoryMethod resolving a proxy without a target throws NullReferenceException instead of a better exception.

For now just throwing notsupportedexception. I guess we _might_ add support for that scenario but we need to find a compelling case for it first. As it stands right now, I have no idea what a good behavior might be