added default build configuration for VS
this will improve first-time VS run experience for people who download the code and try to run it in Visual Studio.
Previously the code would open in default DEBUG configuration or MONO-28 and most likely fail building.

Now the default configuration in NET40-Debug and it should build without issues on any bare machine with VS2010 or 2012